feat: swap default dashboard routes
