+++ src/app/(main)/dashboard/(legacy)/default-v1/_components/chart-area-interactive.tsx
