+++ src/app/(main)/dashboard/(legacy)/default-v1/_components/proposal-sections-table/columns.tsx
