+++ src/app/(main)/dashboard/(legacy)/default-v1/page.tsx
@@ -1,0 +1,14 @@
+import { ChartAreaInteractive } from "./_components/chart-area-interactive";
+import data from "./_components/data.json";
+import { ProposalSectionsTable } from "./_components/proposal-sections-table/table";
+import { SectionCards } from "./_components/section-cards";
+
+export default function Page() {
+  return (
+    <div className="@container/main flex flex-col gap-4 md:gap-6">
+      <SectionCards />
+      <ChartAreaInteractive />
+      <ProposalSectionsTable data={data} />
+    </div>
+  );
+}
