--- src/app/(main)/dashboard/default-v2/page.tsx
@@ -1,13 +1,0 @@
-import { MetricCards } from "./_components/metric-cards";
-import { PerformanceOverview } from "./_components/performance-overview";
-import { SubscriberOverview } from "./_components/subscriber-overview";
-
-export default function Page() {
-  return (
-    <div className="@container/main flex flex-col gap-4 md:gap-6">
-      <MetricCards />
-      <PerformanceOverview />
-      <SubscriberOverview />
-    </div>
-  );
-}
