+++ src/app/(main)/dashboard/default/_components/recent-customers-table/schema.ts
