+++ src/app/(main)/dashboard/default/page.tsx
@@ -1,14 +1,13 @@
-import { ChartAreaInteractive } from "./_components/chart-area-interactive";
-import data from "./_components/data.json";
-import { ProposalSectionsTable } from "./_components/proposal-sections-table/table";
-import { SectionCards } from "./_components/section-cards";
+import { MetricCards } from "./_components/metric-cards";
+import { PerformanceOverview } from "./_components/performance-overview";
+import { SubscriberOverview } from "./_components/subscriber-overview";
 
 export default function Page() {
   return (
     <div className="@container/main flex flex-col gap-4 md:gap-6">
-      <SectionCards />
-      <ChartAreaInteractive />
-      <ProposalSectionsTable data={data} />
+      <MetricCards />
+      <PerformanceOverview />
+      <SubscriberOverview />
     </div>
   );
 }
